--- modules/clear.js
@@ -1,1 +1,0 @@
-export * from './clear.js';
